BlAsyncImageElement class comments, clear forms for new images
Infinite elements may reuse child elements to minimise garbage.  Clear
the form and bitmap in BlFileImageElement and BlUrlImageElement when
specifying the source in case the element is being reused.

BlImageCacheHousekeeper>>initialize: set the currentSize in case the
house keeping process doesn't get a chance to run before images are
added.