BlAsyncElement: Reduce unnecessary work
Extend BlImageCache to answer a BlImageCacheEntry which knows whether
the requested form is the smallest available and/or the original form
(which is always the largest available).  This allows the
BlAsyncImageElement to avoid making requests to the cache which will
answer the same form it already has.

Add updateQueued flag to BlAsyncElement, which can then be used to avoid
submitting unnecessary tasks to the background processing queue.  This
is necessary in particular when the queue has a maxPoolSize > 1, meaning
that multiple tasks for the same element can be executed in parallel
doing the same work.